# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "This is a sample API for demonstration purposes.",
        "version": "3.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Server"
        }
    ],
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User registration",
                "description": "Register a new user.",
                "operationId": "ed42f392010c301892e640067141676e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registered successfully"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User login",
                "operationId": "8cb7742ae98d0990acb2907c4e00bf24",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Logged in successfully"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "419": {
                        "description": "session token has expired"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User logout",
                "operationId": "0d29c740dc06d77602af8095bf4654cc",
                "responses": {
                    "200": {
                        "description": "Logged out successfully"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "419": {
                        "description": "Session token missing"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/google/redirect": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Redirect to Google for login",
                "description": "Redirects the user to Google OAuth 2.0 authentication page",
                "operationId": "f6179c4b563e8988de973b5f045a95d0",
                "responses": {
                    "302": {
                        "description": "Redirect to Google login page"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "oauth2": [
                            "google"
                        ]
                    }
                ]
            }
        },
        "/auth/google/callback": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Callback from Google after login",
                "description": "Handles the callback from Google and authenticates the user",
                "operationId": "fa8298fba53c125aa86525bb1e40f5db",
                "responses": {
                    "200": {
                        "description": "Logged in successfully or user created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "User data",
                                            "type": "object"
                                        },
                                        "token": {
                                            "description": "JWT token",
                                            "type": "string"
                                        },
                                        "message": {
                                            "description": "Login message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Bearer Token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth related operations"
        }
    ]
}